grooming_service/requirements.txt
<<<<<<< Updated upstream
=======
Flask==2.3.0
Flask-CORS==4.0.0
onnxruntime==1.18.1          # use onnxruntime-gpu==1.18.1 if you have CUDA
Pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
numpy==1.26.4
google-generativeai==0.8.0
>>>>>>> Stashed changes
